added button 'remove all'
